requirements.txt
# Requirements file for
#   -collcetMetadata
#   -mergeMetadata
#
# @Author: Uwe Schilling
# @Date: 2016-11-22

psutil
py-cpuinfo
sh
logging
click
